helper/requirements.txt
paho_mqtt==2.1.0
requests
numpy==2.2.5
scipy==1.15.2
matplotlib==3.10.3